+++ src/models/index.js
@@ -5,10 +5,10 @@ const {
 // Bring Models
 const User = require("./user.model.js");
 const Store = require("./store.model");
-const ProductStore = require("./products_stores.model.js");
-const Product = require("./products.model.js");
+const ProductStore = require("./products_store.model.js");
+const Product = require("./product.model.js");
 const Color = require("./color.model.js");
-const ColorTones = require("./color_tones.model.js");
+const ColorTone = require("./color_tone.model.js");
 // Set DB
 // Option 3: Passing parameters separately (other dialects)
 const sequelize = new Sequelize(database.dbName, database.user, database.password, {
@@ -24,26 +24,15 @@ const models = {
   Product: Product(sequelize, DataTypes),
   ProductStore: ProductStore(sequelize, DataTypes),
   Color: Color(sequelize, DataTypes),
-  ColorTones: ColorTones(sequelize, DataTypes),
+  ColorTones: ColorTone(sequelize, DataTypes),
 };
 
 // Define Relationships
-// Modelo 3, el del profe
 Object.keys(models).forEach((key) => {
   if ("associate" in models[key]) {
     models[key].associate(models);
   }
 });
-// Intento 2
-//models.Store.hasMany(models.User, { foreignKey: "id_users", sourceKey: "id_users" });
-//models.User.belongsTo(models.Store, { foreignKey: "id_store", targetId: "id_store" });
-// Intento 1
-//models.Store.hasMany(models.User, { as: "users" });
-//models.Store.hasMany(models.ProductStore, { as: "products_stores" });
-//models.ProductStore.belongsTo(models.Store, { as: "store" });
-//models.ProductStore.belongsTo(models.Product, { as: "products" });
-//models.Product.belongsToMany(models.Color, { through: models.ColorTones });
-//models.Color.belongsToMany(models.Product, { through: models.ColorTones });
 
 // export
 module.exports = { sequelize, models };
